tonghua-project/backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy[asyncio]==2.0.35
aiomysql==0.2.0
alembic==1.13.1
pymysql==1.1.0

# Security
python-jose[cryptography]==3.3.0
cryptography==43.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Redis
redis==5.1.0
aioredis==2.0.1

# HTTP
httpx==0.27.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
